fix: don't create an httpconnectionmanager per route
