+++ pkg/plugins/policies/meshhttproute/plugin/v1alpha1/testdata/basic.listeners.golden.yaml
@@ -32,53 +32,11 @@ resources:
                 route:
                   cluster: backend
                   timeout: 0s
-          statPrefix: backend
-    - filters:
-      - name: envoy.filters.network.http_connection_manager
-        typedConfig:
-          '@type': type.googleapis.com/envoy.extensions.filters.network.http_connection_manager.v3.HttpConnectionManager
-          httpFilters:
-          - name: envoy.filters.http.router
-            typedConfig:
-              '@type': type.googleapis.com/envoy.extensions.filters.http.router.v3.Router
-          routeConfig:
-            name: outbound:backend
-            requestHeadersToAdd:
-            - header:
-                key: x-kuma-tags
-                value: '&kuma.io/protocol=http&&kuma.io/service=frontend&'
-            validateClusters: false
-            virtualHosts:
-            - domains:
-              - '*'
-              name: backend
-              routes:
               - match:
                   prefix: /v2
                 route:
                   cluster: backend-_0_
                   timeout: 0s
-          statPrefix: backend
-    - filters:
-      - name: envoy.filters.network.http_connection_manager
-        typedConfig:
-          '@type': type.googleapis.com/envoy.extensions.filters.network.http_connection_manager.v3.HttpConnectionManager
-          httpFilters:
-          - name: envoy.filters.http.router
-            typedConfig:
-              '@type': type.googleapis.com/envoy.extensions.filters.http.router.v3.Router
-          routeConfig:
-            name: outbound:backend
-            requestHeadersToAdd:
-            - header:
-                key: x-kuma-tags
-                value: '&kuma.io/protocol=http&&kuma.io/service=frontend&'
-            validateClusters: false
-            virtualHosts:
-            - domains:
-              - '*'
-              name: backend
-              routes:
               - match:
                   prefix: /
                 route:
